python/gcp_checkRSSFeeds/requirements.txt
feedparser
google-cloud-firestore